requirements.txt
django==5.0.3
pre-commit==3.6.2
djangorestframework~=3.15.1
reportlab~=4.1.0
pypdf2~=3.0.1
